api/requirements.txt
flask
flask_cors
flask_sqlalchemy
SQLAlchemy-serializer
pysqlite3
requests
pillow
websockets
google
flask_oauthlib